Reorganize welcome window into singleton and module
